Norminette refactoring of interpreter.c
